+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changes
 Unreleased
 ----------
 
+- Add JSON support for COSE. `#109 <https://github.com/dajiaji/python-cwt/pull/109>`__
+- Devite a COSE options parameter into independent parameters. `#109 <https://github.com/dajiaji/python-cwt/pull/109>`__
+- Refine COSE default mode. `#108 <https://github.com/dajiaji/python-cwt/pull/108>`__
 - Refine the order of parameters for CWT functions. `#107 <https://github.com/dajiaji/python-cwt/pull/107>`__
 - Fix example in docstring. `#107 <https://github.com/dajiaji/python-cwt/pull/107>`__
 - Make interface docstring public. `#106 <https://github.com/dajiaji/python-cwt/pull/106>`__